backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
minio
pytest
httpx
redis
biopython
fakeredis
pytest-asyncio
celery
primer3-py
elasticsearch
python-barcode[images]
pyotp
requests
fpdf2
locust
prometheus-client
sentry-sdk
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
slowapi
